Moving assets to a RevenueCat folder for better organizing - helps to know what to delete when upgrading
